requirements.txt
PySimpleGUI==4.60.4
rsa==4.9
pycryptodome==3.17
